♻️ split app and server
